Moved some static functions into their own files
